Adding link to workshop webpage for event-031
In addition to the information on the SORSE page for the communities workshop, an external web page has been set up detailing the full schedule and speaker and session information. Adding a link to this web page to the bottom of the session abstract. The web page also links back to the SORSE page for registration.